Convert entire codebase from double to float precision
- Replace all double with float throughout codebase
- Add f suffix to all floating-point literals
- Define M_PI_F for float precision π constant
- Update all WebAssembly bindings to use float
- Fix scientific notation formatting (e.g., 1.23e-4f)
- Maintain API compatibility with JavaScript (double inputs converted to float)
- Improve WebAssembly performance and memory efficiency
- No backward compatibility cruft - clean conversion

+++ include/ballistics/constants.h
@@ -12,26 +12,26 @@ namespace btk::ballistics
   {
     public:
     // Gravity
-    static constexpr double GRAVITY = 9.80665; // m/s² - standard gravitational acceleration at sea level
+    static constexpr float GRAVITY = 9.80665f; // m/s² - standard gravitational acceleration at sea level
 
     // Atmospheric constants
-    static constexpr double AIR_DENSITY_STANDARD = 1.225; // kg/m³ - standard air density at sea level, 15°C
+    static constexpr float AIR_DENSITY_STANDARD = 1.225f; // kg/m³ - standard air density at sea level, 15°C
 
     // Temperature constants
-    static constexpr double TEMPERATURE_STANDARD_FAHRENHEIT = 288.15; // K - standard temperature at sea level (59°F)
-    static constexpr double TEMPERATURE_STANDARD_KELVIN = 288.15;     // K - ICAO standard temperature at sea level
+    static constexpr float TEMPERATURE_STANDARD_FAHRENHEIT = 288.15f; // K - standard temperature at sea level (59°F)
+    static constexpr float TEMPERATURE_STANDARD_KELVIN = 288.15f;     // K - ICAO standard temperature at sea level
 
     // Pressure constants
-    static constexpr double PRESSURE_STANDARD_PASCALS = 101325.0; // Pa - standard atmospheric pressure at sea level
+    static constexpr float PRESSURE_STANDARD_PASCALS = 101325.0f; // Pa - standard atmospheric pressure at sea level
 
     // Atmospheric lapse rates (dimensionless ratios)
-    static constexpr double TEMPERATURE_LAPSE_RATE = -0.0065; // K/m - temperature lapse rate (troposphere)
-    static constexpr double PRESSURE_SCALE_HEIGHT = 8400.0;   // m - atmospheric scale height for pressure
+    static constexpr float TEMPERATURE_LAPSE_RATE = -0.0065f; // K/m - temperature lapse rate (troposphere)
+    static constexpr float PRESSURE_SCALE_HEIGHT = 8400.0f;   // m - atmospheric scale height for pressure
 
     // Gas constants (dimensionless ratios)
-    static constexpr double GAS_CONSTANT_UNIVERSAL = 8.314; // J/(mol·K) - universal gas constant
-    static constexpr double MOLAR_MASS_DRY_AIR = 0.02897;   // kg/mol - molar mass of dry air
-    static constexpr double HEAT_CAPACITY_RATIO_AIR = 1.4;  // dimensionless - heat capacity ratio for air
+    static constexpr float GAS_CONSTANT_UNIVERSAL = 8.314f; // J/(mol·K) - universal gas constant
+    static constexpr float MOLAR_MASS_DRY_AIR = 0.02897f;   // kg/mol - molar mass of dry air
+    static constexpr float HEAT_CAPACITY_RATIO_AIR = 1.4f;  // dimensionless - heat capacity ratio for air
   };
 
 } // namespace btk::ballistics
