+++ include/physics/atmosphere.h
@@ -23,30 +23,30 @@ namespace btk::physics
      *
      * @param temperature Temperature in K
      * @param altitude Altitude in m
-     * @param humidity Relative humidity (0.0 to 1.0)
+     * @param humidity Relative humidity (0.0f to 1.0f)
      * @param pressure Barometric pressure in Pa (0 for standard pressure at altitude)
      */
-    Atmosphere(double temperature, double altitude, double humidity, double pressure = 0.0);
+    Atmosphere(float temperature, float altitude, float humidity, float pressure = 0.0f);
 
     // Getters (all return SI base units)
-    double getTemperature() const { return temperature_; } // K
-    double getAltitude() const { return altitude_; }       // m
-    double getHumidity() const { return humidity_; }       // 0.0 to 1.0
-    double getPressure() const;                            // Pa
+    float getTemperature() const { return temperature_; } // K
+    float getAltitude() const { return altitude_; }       // m
+    float getHumidity() const { return humidity_; }       // 0.0f to 1.0f
+    float getPressure() const;                            // Pa
 
     /**
      * @brief Calculate air density at current conditions
      *
      * @return Air density in kg/m³
      */
-    double getAirDensity() const;
+    float getAirDensity() const;
 
     /**
      * @brief Calculate speed of sound at current conditions
      *
      * @return Speed of sound in m/s
      */
-    double getSpeedOfSound() const;
+    float getSpeedOfSound() const;
 
     /**
      * @brief Create standard atmosphere at sea level
@@ -56,18 +56,18 @@ namespace btk::physics
     /**
      * @brief Create atmosphere for given altitude with standard lapse rates
      */
-    static Atmosphere atAltitude(double altitude); // altitude in m
+    static Atmosphere atAltitude(float altitude); // altitude in m
 
     private:
-    double temperature_; // K
-    double altitude_;    // m
-    double humidity_;    // 0.0 to 1.0
-    double pressure_;    // Pa
+    float temperature_; // K
+    float altitude_;    // m
+    float humidity_;    // 0.0f to 1.0f
+    float pressure_;    // Pa
 
     /**
      * @brief Calculate standard pressure for given altitude
      */
-    double calculateStandardPressure(double altitude) const; // altitude in m, returns Pa
+    float calculateStandardPressure(float altitude) const; // altitude in m, returns Pa
   };
 
 } // namespace btk::physics
